requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.0
plotly==5.19.0
python-dotenv==1.0.1
psycopg2-binary
openai
sentence-transformers
replit